==> YK1 Burned Processing/6 ameriflux processing.Rmd <==
---
title: "Prep for Ameriflux - YK1 Burned"
author: "Kyle Arndt"
date: "2024-02-15"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = 'C:/Users/karndt.WHRC/Desktop/sites/YKD/Ameriflux YK/burned/') #use this to set your directory
```

Libraries
```{r}
rm(list = ls())
library(data.table)
library(ggplot2)
library(zoo)

Sys.setenv(TZ = 'UTC')
```

Load Data
```{r}
df = fread(input = "./ykd_burned_gapfilled_clean_2019_2022.csv")
```

Format the time stamps at start and end
```{r}
TIMESTAMP_END = df$ts

yeare = format(TIMESTAMP_END,"%Y")
monthe = format(TIMESTAMP_END,"%m")
daye = format(TIMESTAMP_END,"%d")
houre = format(TIMESTAMP_END,"%H")
mine = format(TIMESTAMP_END,"%M")

TIMESTAMP_END = paste(yeare,monthe,daye,houre,mine,sep = '')

TIMESTAMP_START = df$ts - 30*60

years = format(TIMESTAMP_START,"%Y")
months = format(TIMESTAMP_START,"%m")
days = format(TIMESTAMP_START,"%d")
hours = format(TIMESTAMP_START,"%H")
mins = format(TIMESTAMP_START,"%M")

TIMESTAMP_START = paste(years,months,days,hours,mins,sep = '')
rm(daye,days,mine,mins,houre,hours,monthe,months,yeare,years)
```


Format and name data
```{r}
#gas concentrations and variance ################################
CO2 = df$co2_mole_fraction
CH4 = ifelse(df$ch4_mole_fraction<0,NA,df$ch4_mole_fraction*1000)
H2O = df$h2o_mole_fraction

CO2_MIXING_RATIO = df$co2_mixing_ratio
CH4_MIXING_RATIO = ifelse(df$ch4_mixing_ratio<0,NA,df$ch4_mixing_ratio*1000)
#H2O mole fraction and H2O mixing ratio are identical

#standard deviation of co2 wet fraction. (Standard deviation is sqrt of the variance)
# CO2_SIGMA = sqrt(df$co2_var)
# CH4_SIGMA = sqrt(df$ch4_var*1000)
# H2O_SIGMA = sqrt(df$h2o_var/1000)

#Gas Fluxes ######################
#CO2 flux is umol m-2 s-1
FC   = df$co2_flux.c
FC_F = ifelse(is.na(df$co2_flux.c),df$rfnee,df$co2_flux.c)
FCH4 = df$ch4_flux.c*1000
FCH4_F = ifelse(is.na(df$ch4_flux.c),df$rfch4,df$ch4_flux.c)*1000
FH2O = df$h2o_flux

#GPP_F   = df$GPP_f
GPP     = ifelse(is.na(df$co2_flux.c),NA,df$GPP_f)
GPP_F   = df$GPP_f
#RECO_F  = df$Reco
RECO     = ifelse(is.na(df$co2_flux.c),NA,df$Reco)
RECO_F = df$Reco

#energy fluxes ###########################
G_1_1_1 = df$SHF_1_1_1.c
G_2_1_1 = df$SHF_2_1_1.c
G_3_1_1 = ifelse(df$SHF_3_1_1.c > 125,NA,df$SHF_3_1_1.c)

H       = df$H.c
LE      = df$LE.c

#meteorology and weather #################################
PA      = df$air_pressure/1000  #air pressure in kPA
TA      = df$TA_1_1_1.c    # Air T in Deg C
SVP     = 610.7*10^((7.5*TA)/(237.3+TA)) #NOT INCLUDED IN FINAL
RH      = df$RH_1_1_1.c   #RH in %
T_SONIC = df$sonic_temperature-273.15 #sonic temp in Deg C
VPD     = ((100 - RH)/100)*SVP/100 #VPD in hPA
P       = df$P_RAIN_1_1_1*1000 # rain in mm

#Radiation #####################################
ALB = df$ALB_1_1_1.c*100 #albedo in %

LW_IN  = df$LWIN_1_1_1.c  # Wm-2
LW_OUT = df$LWOUT_1_1_1.c # Wm-2
SW_IN  = df$SWIN_1_1_1.c   # Wm-2
SW_OUT = df$SWOUT_1_1_1.c # Wm-2

NETRAD = df$RN_1_1_1.c     # Wm-2

PPFD_IN = df$PPFD_1_1_1.c   # par in umol m-2 s-1

#Soil factors ###################################
#all swc in %
SWC_1_1_1 = df$SWC_1_1_1.c
SWC_2_1_1 = df$SWC_2_1_1.c
SWC_3_1_1 = df$SWC_3_1_1.c

#all soil temps deg C
TS_1_1_1 = df$TS_1_1_1.c
TS_2_1_1 = df$TS_2_1_1.c
TS_3_1_1 = df$TS_3_1_1.c

#wind and turbulence parameters ##############################
MO_LENGTH = df$L
ZL        = df$`(z-d)/L`
TAU       = df$Tau

U_SIGMA = sqrt(df$u_var)
V_SIGMA = sqrt(df$v_var)
W_SIGMA = sqrt(df$w_var)

USTAR = df$`u*`

WD     = df$wind_dir
WS     = df$wind_speed
WS_MAX = df$max_wind_speed

#Quality flags from EP (Mauder and Foken) ################
FC_SSITC_TEST   = df$qc_co2_flux
FCH4_SSITC_TEST = df$qc_ch4_flux
H_SSITC_TEST    = df$qc_H
LE_SSITC_TEST   = df$qc_LE
TAU_SSITC_TEST  = df$qc_Tau
```


Plots
```{r}
plot(CO2) #umol
plot(CH4) #nmol
plot(H2O) #mmol
```

```{r}
plot(CO2_MIXING_RATIO) #umol
plot(CH4_MIXING_RATIO) #nmol
```

```{r}
# plot(CO2_SIGMA)
# plot(CH4_SIGMA) 
# plot(H2O_SIGMA) 
```

```{r}
plot(FC)
plot(FC_F)
plot(FCH4)
plot(FCH4_F)
plot(FH2O)
plot(GPP_F)
plot(GPP)
plot(RECO_F)
plot(RECO)
```

```{r}
plot(G_1_1_1)
plot(G_2_1_1)
plot(G_3_1_1)

plot(H)
plot(LE)
```


```{r}
plot(PA)
plot(RH)
plot(T_SONIC)
plot(TA)   
plot(VPD) 
plot(P)
```

```{r}
plot(ALB)
plot(LW_IN)
plot(LW_OUT)
plot(SW_IN)
plot(SW_OUT)
plot(NETRAD)
plot(PPFD_IN)
```


```{r}
plot(SWC_1_1_1)
plot(SWC_2_1_1)
plot(SWC_3_1_1)

plot(TS_1_1_1)
plot(TS_2_1_1)
plot(TS_3_1_1)
```

```{r}
plot(MO_LENGTH)
plot(ZL)
plot(TAU)
plot(U_SIGMA)
plot(V_SIGMA)
plot(W_SIGMA)
plot(USTAR)
plot(WD)
plot(WS)
plot(WS_MAX)
```

```{r}
plot(FC_SSITC_TEST)
plot(FCH4_SSITC_TEST)
plot(H_SSITC_TEST)
plot(LE_SSITC_TEST)
plot(TAU_SSITC_TEST)
```


```{r}
US_YK1 = data.frame(TIMESTAMP_START,TIMESTAMP_END,
  CO2,CH4,H2O
,CO2_MIXING_RATIO,CH4_MIXING_RATIO
#,CO2_SIGMA,CH4_SIGMA ,H2O_SIGMA 
,FC,FCH4,FH2O
,FC_F,FCH4_F,GPP,GPP_F,RECO_F,RECO
,G_1_1_1,G_2_1_1,G_3_1_1
,H
,LE
,PA
,RH
,T_SONIC
,TA   
,VPD 
,P
,ALB
,LW_IN
,LW_OUT
,SW_IN
,SW_OUT
,NETRAD
,PPFD_IN
,SWC_1_1_1
,SWC_2_1_1
,SWC_3_1_1
,TS_1_1_1
,TS_2_1_1
,TS_3_1_1
,MO_LENGTH,ZL,TAU,U_SIGMA,V_SIGMA,W_SIGMA,USTAR,WD,WS,WS_MAX
,FC_SSITC_TEST,FCH4_SSITC_TEST,H_SSITC_TEST,LE_SSITC_TEST,TAU_SSITC_TEST)
```


```{r}
siteID = "US-YK1"
resolution = 'HH'
tsstart = US_YK1$TIMESTAMP_START[1]
tsend   = US_YK1$TIMESTAMP_END[nrow(US_YK1)]

filename = paste(siteID,resolution,tsstart,tsend,sep = '_')
filename = paste("./",filename,".csv",sep = "")

write.csv(x = US_YK1,file = filename,row.names = F,na = "-9999",quote = F)
```
